requirements.txt
transformers
torch
pillow
streamlit
